bench: make the backends precision-agnostic, and measure that float loses
BackendConfig.hpp has always claimed the field element type is a one-line
swap. It nearly was -- three std::max(double&, float) calls in adaptTimeStep,
and kbench's snapshot buffers hardcoded std::vector<double>. Both fixed, so
`using Field = GridField<float>` now compiles and runs on the serial backend.
No behaviour change at double; the casts are no-ops there.

The point was to size single precision as an optimisation, since it halves
memory traffic and would remove CUDA's 1/64 FP64 penalty. On the CPU it is a
LOSS. Whole application, production config, paired, 8 pairs:

  serial double  0.6644 s/step
  serial float   0.7242 s/step     0.941x -- 0 of 8 pairs faster

Per kernel, 360x120x60 RoundedCorner, Re = 10000:

  kernel                     double    float   ratio
  updateVelocities             5.03     7.18   1.43x slower
  solvePressurePoisson        47.82    67.66   1.41x slower
  buildPressureSource         15.41    17.18   1.11x slower
  computeAccelerations        78.53    76.76   0.98x
  adaptTimeStep                1.86     1.75   0.94x faster

The pattern is exact and explains itself. Only the STORAGE changed; every
kernel still computes in double, so each field access pays a float->double
conversion and each store the reverse. The kernels that lose worst are the ones
with many accesses and cheap arithmetic -- solvePressurePoisson reads seven
neighbours and writes one for about 13 FLOPs, so conversions dominate the
bandwidth they save. The single winner, adaptTimeStep, is read-only with
trivial math: pure streaming, which is the only case where halving the bytes
pays. computeAccelerations is a wash because exp() stays double and dominates
regardless.

So storage-only precision reduction is the wrong shape of change. A real win
needs end-to-end float -- float locals, expf, float scratch -- which removes the
conversions, makes the transcendentals cheaper, and doubles SIMD width. On x86
that is bounded (FP32 and FP64 differ only by vector width, so maybe 1.4-1.6x);
on CUDA it is the whole ballgame, since GA107 runs FP64 at 1/64 of FP32 and the
CPU measured here has 183.9 GFLOP/s FP64 against the GPU's ~77.

Kept because the swap now works and the measurement is reproducible by anyone
editing one line. Default stays double. ctest 5/5.

+++ src/backends/openmp/Physics.cpp
@@ -825,9 +825,9 @@ void adaptTimeStep(SimState& s)
     for (int i = 1; i <= s.numCellsXm1; ++i)
         for (int j = s.jLow[i]+1; j <= s.jHigh[i]-1; ++j)
             for (int k = 1; k <= cfg.numCellsZ; ++k) {
-                uMax = std::max(uMax, std::abs(s.velX(i,j,k)));
-                vMax = std::max(vMax, std::abs(s.velY(i,j,k)));
-                wMax = std::max(wMax, std::abs(s.velZ(i,j,k)));
+                uMax = std::max(uMax, static_cast<double>(std::abs(s.velX(i,j,k))));
+                vMax = std::max(vMax, static_cast<double>(std::abs(s.velY(i,j,k))));
+                wMax = std::max(wMax, static_cast<double>(std::abs(s.velZ(i,j,k))));
             }
 
     const double reForDiff = (cfg.hyperViscousStart == 0) ? cfg.reynoldsNumber : cfg.hyperViscousRe;
